// frontend/ticket-booking-web/src/app/api/model/event-category.ts
/**
 * Ticket Booking Platform API - v1
 *
 * Contact: m1musbal@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const EventCategory = {
    Concert: 'Concert',
    Theater: 'Theater',
    Conference: 'Conference',
    Other: 'Other'
} as const;
export type EventCategory = typeof EventCategory[keyof typeof EventCategory];