// frontend/ticket-booking-web/src/app/api/model/event-status.ts
/**
 * Ticket Booking Platform API - v1
 *
 * Contact: m1musbal@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const EventStatus = {
    Draft: 'Draft',
    OnSale: 'OnSale',
    SoldOut: 'SoldOut',
    Cancelled: 'Cancelled',
    Completed: 'Completed'
} as const;
export type EventStatus = typeof EventStatus[keyof typeof EventStatus];